API

DynamicSumTypes.@sumtype — Macro

@sumtype SumTypeName(Types) [<: AbstractType]

The macro creates a sumtypes composed by the given types. It optionally accept also an abstract supertype.

Example

julia> using DynamicSumTypes

julia> struct A x::Int end;

julia> struct B end;

julia> @sumtype AB(A, B)

DynamicSumTypes.variant — Function

variant(inst)

Returns the variant enclosed in the sum type.

Example

julia> using DynamicSumTypes

julia> struct A x::Int end;

julia> struct B end;

julia> @sumtype AB(A, B)

julia> a = AB(A(0))
AB'.A(0)

julia> variant(a)
A(0)

DynamicSumTypes.allvariants — Function

allvariants(SumType)

Returns all the enclosed variants types in the sum type in a namedtuple.

Example

julia> using DynamicSumTypes

julia> struct A x::Int end;

julia> struct B end;

julia> @sumtype AB(A, B)

julia> allvariants(AB)
(A = A, B = B)

DynamicSumTypes.variantof — Function

variantof(inst)

Returns the type of the variant enclosed in the sum type.

DynamicSumTypes.is_sumtype — Function

is_sumtype(T)

Returns true if the type is a sum type otherwise returns false.