API

MixedStructTypes.@compact_structs — Macro

@compact_structs(type_definition, structs_definitions)

This macro allows to combine multiple types in a single one. This version has been built to yield a performance almost identical to having just one type.

Example

julia> @compact_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> a = A(1)
A(1)::AB

julia> a.x
1

MixedStructTypes.@sum_structs — Macro

@sum_structs(type_definition, structs_definitions)

This macro allows to combine multiple types in a single one. While its usage is equivalent to @compact_structs, this version consumes less memory at the cost of being slower.

Example

julia> @sum_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> a = A(1)
A(1)::AB

julia> a.x
1

MixedStructTypes.@dispatch — Macro

@dispatch(function_definition)

This macro allows to dispatch on types created by @compact_structs or @sum_structs. Notice that this only works when the kinds in the macro are not wrapped by any type containing them.

Example

julia> @compact_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> @dispatch f(::A) = 1;

julia> @dispatch f(::B) = 2;

julia> @dispatch f(::Vector{B}) = 3; # this doesn't work
ERROR: UndefVarError: `B` not defined
Stacktrace:
 [1] top-level scope
   @ REPL[7]:1

julia> f(A(0))
1

julia> f(B(0))
2

MixedStructTypes.kindof — Function

kindof(instance)

Return a symbol representing the conceptual type of an instance:

julia> @compact_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> a = A(1);

julia> kindof(a)
:A

MixedStructTypes.allkinds — Function

allkinds(type)

Return a Tuple containing all kinds associated with the overarching type defined with @compact_structs or @sum_structs:

julia> @compact_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> allkinds(AB)
(:A, :B)

MixedStructTypes.kindconstructor — Function

kindconstructor(instance)

Return the constructor of an instance:

julia> @compact_structs AB begin
           struct A x::Int end
           struct B y::Int end
       end

julia> a = A(1);

julia> kindconstructor(a)
A